putter react-table useTable inn i OppgaverContext, bruker usePagination
